requirements.txt
absl-py==0.13.0
asgiref==3.2.10
astunparse==1.6.3
backports.pbkdf2==0.1
beautifulsoup4==4.9.3
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.6.20
charset-normalizer==2.0.4
clang==5.0
cramjam==2.4.0
cvlib==0.2.6
cycler==0.10.0
Django==3.1.2
django-bootstrap4==2.3.1
fastparquet==0.7.1
flatbuffers==1.12
fsspec==2021.10.1
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
idna==3.2
ImageHash==4.1.0
imageio==2.9.0
importlib-metadata==2.0.0
imutils==0.5.4
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.4
matplotlib==3.3.2
Naked==0.1.31
numpy==1.19.2
oauthlib==3.1.1
opencv-python==4.4.0.44
opt-einsum==3.3.0
pandas==1.3.4
piexif==1.1.3
Pillow==8.0.1
progressbar==2.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.4
pyparsing==2.4.7
PySimpleGUI==4.30.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.3
shellescape==3.8.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
thrift==0.15.0
tk==0.1.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.0
